+++ include/drivers/HMC5883L.hpp
