+++ src/drivers/HMC5883L.cpp
@@ -21,7 +21,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-#include "HMC5883L.hpp"
+#include "drivers/HMC5883L.hpp"
 #include "i2c_helpers.hpp"
 bool HMC5883L::begin() {
   // Wire.begin();
